Added ExportComponentMixin and added export component functionality to Export Currently Selec ted button in home side bar. Co-authored by: Julia Bakerink juliabakerink@gmail.com Co-authored by: Megan Nadkarni megan.nadkarni@gmail.com
