--- Test3/package.json
@@ -1,48 +1,0 @@
-{
-	"name": "vue-boiler-plate-routes",
-	"version": "0.1.0",
-	"private": true,
-	"scripts": {
-		"start": "vue-cli-service serve",
-		"build": "vue-cli-service build",
-		"lint": "vue-cli-service lint"
-	},
-	"dependencies": {
-		"vue": "^3.2.26",
-		"vue-router": "^4.0.12",
-		"vuex": "^4.0.2"
-	},
-	"devDependencies": {
-		"@vue/cli-plugin-babel": "~4.5.0",
-		"@vue/cli-plugin-eslint": "~4.5.0",
-		"@vue/cli-service": "~4.5.0",
-		"babel-eslint": "^10.0.1",
-		"eslint": "^6.7.2",
-		"eslint-plugin-vue": "^7.0.0-0",
-		"@vue/compiler-sfc": "^3.0.0-0"
-	},
-	"eslintConfig": {
-		"root": true,
-		"env": {
-			"node": true
-		},
-		"extends": [
-			"plugin:vue/essential",
-			"eslint:recommended"
-		],
-		"rules": {},
-		"parserOptions": {
-			"parser": "babel-eslint"
-		}
-	},
-	"postcss": {
-		"plugins": {
-			"autoprefixer": {}
-		}
-	},
-	"browserslist": [
-		"> 1%",
-		"last 2 versions",
-		"not ie <= 8"
-	]
-}
