--- Test3/public/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width,initial-scale=1.0">
-	<link rel="icon" href="<%= BASE_URL %>favicon.ico">
-	<title>New Vue Project</title>
-</head>
-
-<body>
-	<noscript>
-		<strong>We're sorry but vue-boiler-plate-routes doesn't work properly without JavaScript enabled. Please enable it
-      to continue.</strong>
-	</noscript>
-	<div id="app"></div>
-	<!-- built files will be auto injected -->
-</body>
-
-</html>
