--- Test3/src/components/Test.vue
@@ -1,17 +1,0 @@
-<template>
-	<div>
-    <div></div>
-	</div>
-</template>
-
-<script>
-
-export default {
-  name: 'Test',
-  components: {
-  },
-};
-</script>
-
-<style scoped>
-</style>
