--- Test3/src/views/HomeView.vue
@@ -1,20 +1,0 @@
-<template>
-	<div>
-	</div>
-</template>
-
-<script>
-import Test from '@/components/Test.vue';
-import Test2 from '@/components/Test2.vue';
-
-export default {
-  name: 'HomeView',
-  components: {
-    Test,
-    Test2,
-  }
-};
-</script>
-
-<style scoped>
-</style>
