+++ TestExport3/Test.vue
@@ -5,10 +5,12 @@
 </template>
 
 <script>
+import ButtonTest from '@/components/ButtonTest.vue';
 
 export default {
   name: 'Test',
   components: {
+    ButtonTest,
   },
 };
 </script>
